Deprecated `Project.get_files` (fixes #101)
`Project.get_files` is very similar to `Package.get_list` and `Package.get_files`;
therefore it is now marked as deprecated and calls the other methods.

To work around inconsistent parameter validation in the OBS API, which caused #101
in the first place, the `Package` extension is now "sanitizing" the query
parameters to the liking of build service.