Enhanced usability and reliability for `HttpSignatureAuth`
The old implementation made the incorrect assumption, that reading a private key with `ssh-keygen`
would imply getting it from the agent, if it is available. This caused an additional prompt for
the passphrase in interactive sessions. Instead, `is_ssh_key_readable` attempts to sign a dummy
message.

This ensures, that an exception still gets raised early and that no superfluous prompt is shown.

In addition, any SSH command is terminated after a timeout (currently 10 seconds) to ensure that no
applications blocks indefinitely due to a passphrase prompt, which might still happen.